requirements.txt
Flask==3.0.0
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
grpcio>=1.60.0
tiktoken==0.5.1
python-dotenv==1.0.0
requests==2.31.0
